feat: extract antd/arco/splitChunks logics into react plugin
